restart_interval converted from float to int (Error with real version 4.2.2)

+++ wrf4g/config.py
@@ -157,7 +157,7 @@ class SanityCheck():
                         simult_interval = simult_length = end_date - start_date
                     # Defining restart_interval
                     # To avoid chunk restart we add 1 hour to restart_interval variable
-                    restart_interval = timedelta_total_seconds( chunk_size ) / 60
+                    restart_interval = int(timedelta_total_seconds( chunk_size ) / 60)
                     if self.cfg[ section ].get( 'chunk_restart' ) or self.cfg[ section ].get( 'chunk_restart' ) == 'no' :
                         restart_interval = restart_interval + 3600
                     self.cfg_final[ section ][ 'date_time' ].append( [ start_date, end_date,
